Refactoring os commands wrapper

+++ src/Helpers/OsCommandsWrapper.php
@@ -10,7 +10,10 @@
 namespace JustMisha\MultiRunner\Helpers;
 
 use Error;
+use FilesystemIterator;
 use InvalidArgumentException;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
 use RuntimeException;
 
 /**
@@ -41,6 +44,9 @@ class OsCommandsWrapper
      */
     public function removeDirRecursive(string $dir): void
     {
+        if (!file_exists($dir) || !is_dir($dir)) {
+            return;
+        }
         if ($this->isWindows()) {
             exec(sprintf("rd /s /q %s", escapeshellarg($dir)));
             return;
@@ -176,5 +182,30 @@ class OsCommandsWrapper
             throw new RuntimeException("There was error while escaping string " . $value);
         }
         return $result;
+    }
+
+
+    /**
+     * Empty the contents of the folder.
+     *
+     * @param string $folder A full path to the folder to empty.
+     * @return void
+     */
+    public function clearFolder(string $folder): void
+    {
+        if (!file_exists($folder) || !is_dir($folder)) {
+            return;
+        }
+        $files = new RecursiveIteratorIterator(
+            new RecursiveDirectoryIterator($folder, FilesystemIterator::SKIP_DOTS),
+            RecursiveIteratorIterator::CHILD_FIRST
+        );
+        foreach ($files as $file) {
+            if ($file->isDir() === true) {
+                rmdir($file->getPathName());
+            } elseif (($file->isFile() === true) || ($file->isLink() === true)) {
+                unlink($file->getPathname());
+            }
+        }
     }
 }
